2_Resume_AI_Bot/requirements.txt
Flask
pdfplumber
docx2txt
langchain
langchain-google-genai
python-dotenv
langchain_community
